separate image rendering layer
