+++ src/PotoCanvas.css
@@ -1,5 +1,2 @@
 .PotoCanvas {
 }
-  .PotoCanvas-canvas {
-    padding: var(--app-space);
-  }
