+++ src/PotoCanvasImage.css
@@ -1,0 +1,3 @@
+.PotoCanvasImage {
+  padding: var(--app-space);
+}
